always use "now" as model timestamp
Originally I was using the creation timestamp sent with the resource
event from Kubernetes as the Zenoss model timestamp on resource add
events, and the deletion timestamp for delete update. This ends up not
being the best way to do things. The "now" timestamp should be used in
all of these cases so that the Zenoss entities end up existing for the
same time period that the metrics for the entities do.

In the future we'll be adding more metadata fields to capture things
like creation timestamp on the model even if we don't use them as the
model's timestamp.

Fixes ZPS-6222.